+++ app/controllers/admin/businesses/hiring_invoice_requests_controller.rb
@@ -1,0 +1,13 @@
+module Admin
+  module Businesses
+    class HiringInvoiceRequestsController < ApplicationController
+      def index
+        @forms = ::Businesses::HiringInvoiceRequest.includes(:business).order(created_at: :desc)
+      end
+
+      def show
+        @form = ::Businesses::HiringInvoiceRequest.find(params[:id])
+      end
+    end
+  end
+end