+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### March
 
+* March 21 - Added business new hire forms and admin new hire forms #818 @RomanTurner @MikeRayUX
 * March 21 - Announce specialties to developers #817 @sarvaiyanidhi
 * March 20 - Update developer query to not duplicate records with multiple matching specialties #822 @bschrag620
 * March 18 - Add developer specialties #803